+++ Tests/Editor/Container/InstanceBindingTest.cs
@@ -49,5 +49,51 @@ namespace TheRealIronDuck.Ducktion.Editor.Tests.Editor.Container
             
             Assert.AreSame(gObj, container.Resolve<GameObject>());
         }
+
+        [Test]
+        public void ItCanOverrideAServiceWithASpecificInstance()
+        {
+            container.Register<SimpleService>();
+            
+            var old = container.Resolve<SimpleService>();
+            
+            var service = new SimpleService();
+            container.Override<SimpleService>(service);
+            
+            var current = container.Resolve<SimpleService>();
+            
+            Assert.AreSame(service, current);
+            Assert.AreNotSame(old, current);
+        }
+
+        [Test]
+        public void ItCanOverrideAnExistingInstance()
+        {
+            var serviceA = new SimpleService();
+            var serviceB = new SimpleService();
+            
+            container.Register<SimpleService>(serviceA);
+            container.Override<SimpleService>(serviceB);
+            
+            var current = container.Resolve<SimpleService>();
+            
+            Assert.AreSame(serviceB, current);
+        }
+
+        [Test]
+        public void ItCanOverrideUsingTypeParameters()
+        {
+            container.Register<SimpleService>();
+            
+            var old = container.Resolve<SimpleService>();
+            
+            var service = new SimpleService();
+            container.Override(typeof(SimpleService), service);
+            
+            var current = container.Resolve<SimpleService>();
+            
+            Assert.AreSame(service, current);
+            Assert.AreNotSame(old, current);
+        }
     }
 }