+++ CHANGELOG.md
@@ -16,6 +16,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Added method `Register<T>(Instance)` to register a type with a singleton instance
   - Added method `Register(Type, Instance)` to register a type with a singleton instance
   - Added method `Override<TKey, TService>` to override any service with a specific implementation
+  - Added method `Override(KeyType, ServiceType)` to override any service with a specific implementation
+  - Added method `Override<T>(Instance)` to override any service with a specific singleton instance
+  - Added method `Override(Type, Instance)` to override any service with a specific singleton instance
   - Added method `Resolve<T>` to resolve a service
   - Added method `Resolve(Type)` to resolve a service
   - Added method `Clear` to remove any registered service and singleton instance